Rack packaging doesn't fully override rails logging Refactored platform hooks in to the Rails application by using the correct Railtie approach Refactored Ruby code and initialization to separate classes into separate files and clean up code
